requirements.txt
streamlit==1.39.0
google-generativeai
python-dotenv
Pillow==10.3.0
pandas==2.2.2
pymupdf

